requirements.txt
Pillow
exifread
piexif
PyMuPDF
pdfminer.six
python-magic
numpy
scipy
matplotlib
stegano
